+++ Changes.md
@@ -1,3 +1,5 @@
+* #1318 - Update score board text on the dynamic scoreboard when it changes
+* #1310 - Add a clock to the scoreboard display
 * Add edit link to regular match play runs report
 * Display run name on unverified runs page
 * #1317 - handle no scores when creating top performance score reports